backend/requirements.txt

python-jose[cryptography]  
passlib[bcrypt]            
python-multipart           


sqlalchemy                 
alembic                    
psycopg2-binary           


gunicorn                  


pytest
pytest-asyncio


fastapi
uvicorn[standard]
python-dotenv
requests
pydantic